rename services navigator
